refactor: add new `UPDATE_INPUT_STEP` action
This is definitely not the best solution since it complicates the set of
actions that can be applied to our node graph. However, a better
solution requires larger refactors to shift `GraphContext` higher up in
the component hierarchy (to the form level).

The important thing to realise here is that our form will mutate the
state of our graph via `required_inputs`, therefore, we should have
better context manager to reflect this in our components.