fix: unexpected node graph editor reload when adding required inputs (#17)
* fix: remove `userInputNode` from testcase component key

A testcase is uniquely identified **solely** by its `id`.
`userInputNode` should not determine whether the node graph editor
should be rerendered/reloaded (unmount + mount).

* refactor: add new `UPDATE_INPUT_STEP` action

This is definitely not the best solution since it complicates the set of
actions that can be applied to our node graph. However, a better
solution requires larger refactors to shift `GraphContext` higher up in
the component hierarchy (to the form level).

The important thing to realise here is that our form will mutate the
state of our graph via `required_inputs`, therefore, we should have
better context manager to reflect this in our components.

* refactor: add `input` prop `NodeGraph`

This is to separate concern/logic `userInputNode` from the other nodes
in a testcase **in edit mode**.

* fix: update RF graph based on `nodeData`

* chore: remove deprecated components

* refactor: rename node graph actions

* refactor: migrate to new graph actions

* fix: missing `asChild` for socket delete button

* refactor: assume `UserInputStep` is always `id: 0`

* fix: change node socket delete button type to `button`